Change database integer sizes and ensure no negative transfers can occur

+++ migrations/20220331171509_create_accounts_transfers_transactions.up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE accounts (
     id SERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    balance INTEGER NOT NULL,
-    owner_id INTEGER NOT NULL,
+    balance BIGINT NOT NULL,
+    owner_id INT NOT NULL,
 
     CONSTRAINT fk_owner_id
         FOREIGN KEY (owner_id)
@@ -12,9 +12,9 @@ SELECT setval('accounts_id_seq', 999);
 
 CREATE TABLE transfers (
     id SERIAL PRIMARY KEY,
-    from_account INTEGER NOT NULL,
-    to_account INTEGER NOT NULL,
-    amount INTEGER NOT NULL,
+    from_account INT NOT NULL,
+    to_account INT NOT NULL,
+    amount BIGINT NOT NULL,
     created_at timestamptz NOT NULL DEFAULT CURRENT_TIMESTAMP,
 
     CONSTRAINT fk_from_account
